refactor(wire-protocol): flatten wire protocol handler hierarchy
See the ticket for more details, but this generally removes the
class-based module design for wire protocol handling, removing
needless allocations, and focusing more on the meat of the methods
hung off those classes.

NODE-1835
